Restructure app folder layout for better organization.
Reorganized the folder structure to group related pages and layouts under `(dashboard)` and `(root)` directories. Added new layouts for both root and dashboard sections, improving component hierarchy and consistency. Introduced a loading component for global use.
