+++ app/(dashboard)/dashboard/analytics/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const Page = () => {
+    return (
+        <div>Analytics</div>
+    )
+}
+export default Page
