+++ app/(dashboard)/layout.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Layout = ({children}: { children: React.ReactNode }) => {
+    return (
+        <div>
+            <p className="text-3xl">Dashboard</p>
+            {children}
+        </div>
+    )
+}
+
+export default Layout
