+++ app/(root)/about/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const Page = () => {
+    return (
+        <div>About</div>
+    )
+}
+export default Page
