--- app/about/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const Page = () => {
-    return (
-        <div>About</div>
-    )
-}
-export default Page
