--- app/dashboard/analytics/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const Page = () => {
-    return (
-        <div>Analytics</div>
-    )
-}
-export default Page
