--- app/dashboard/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const Page = () => {
-    return (
-        <div>Dashboard</div>
-    )
-}
-export default Page
