+++ app/loading.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const Loading = () => {
+    return (
+        <div>----------------------------------------------------</div>
+    )
+}
+export default Loading
